WIP refactor and tidyup
